added newsView file for SOLID
